Documents the feature flags service
