+++ lib/arcane.dart
@@ -1,6 +1,6 @@
 import "package:arcane_framework/src/authentication/authentication_service.dart";
+import "package:arcane_framework/src/feature_flags/feature_flags_service.dart";
 import "package:arcane_framework/src/providers/service_provider.dart";
-import "package:arcane_framework/src/services/feature_flags.dart";
 import "package:arcane_framework/src/services/id.dart";
 import "package:arcane_framework/src/services/logger.dart";
 import "package:arcane_framework/src/services/theme.dart";
