+++ lib/arcane_framework.dart
@@ -4,9 +4,9 @@ export "package:app_tracking_transparency/app_tracking_transparency.dart";
 export "package:arcane_framework/arcane.dart";
 export "package:arcane_framework/src/arcane_app.dart";
 export "package:arcane_framework/src/authentication/authentication_service.dart";
+export "package:arcane_framework/src/feature_flags/feature_flags_service.dart";
 export "package:arcane_framework/src/providers/environment_provider.dart";
 export "package:arcane_framework/src/providers/service_provider.dart";
-export "package:arcane_framework/src/services/feature_flags.dart";
 export "package:arcane_framework/src/services/id.dart";
 export "package:arcane_framework/src/services/logger.dart";
 export "package:arcane_framework/src/services/theme.dart";
