+++ lib/src/providers/service_provider.dart
@@ -2,9 +2,30 @@ import "package:arcane_framework/arcane_framework.dart";
 import "package:collection/collection.dart";
 import "package:flutter/widgets.dart";
 
+/// A provider that makes a list of `ArcaneService` instances available to the widget tree.
+///
+/// This class extends `InheritedNotifier` and allows `ArcaneService` instances to be
+/// accessed throughout the widget tree by descendant widgets. It should be used to
+/// provide service instances that are shared across the application.
+///
+/// Example:
+/// ```dart
+/// ArcaneServiceProvider(
+///   serviceInstances: [myService],
+///   child: MyApp(),
+/// );
+/// ```
+/// To access the provided services:
+/// ```dart
+/// final provider = ArcaneServiceProvider.of(context);
+/// ```
 class ArcaneServiceProvider extends InheritedNotifier {
+  /// A list of `ArcaneService` instances available through the provider.
   final List<ArcaneService> serviceInstances;
 
+  /// Creates an `ArcaneServiceProvider` that provides [serviceInstances] to the widget tree.
+  ///
+  /// The [child] widget will be the root of the widget subtree that has access to the services.
   @override
   const ArcaneServiceProvider({
     required this.serviceInstances,
@@ -12,11 +33,25 @@ class ArcaneServiceProvider extends InheritedNotifier {
     super.key,
   });
 
+  /// Determines whether the widget should notify its dependents.
+  ///
+  /// This always returns `true`, meaning dependents will always be notified
+  /// when this widget is rebuilt.
   @override
   bool updateShouldNotify(ArcaneServiceProvider oldWidget) {
     return true;
   }
 
+  /// Retrieves the nearest `ArcaneServiceProvider` in the widget tree.
+  ///
+  /// This method is used to access the `ArcaneServiceProvider` and its provided services
+  /// from any descendant widget. It throws an exception if no `ArcaneServiceProvider`
+  /// is found in the widget tree.
+  ///
+  /// Example:
+  /// ```dart
+  /// final provider = ArcaneServiceProvider.of(context);
+  /// ```
   static ArcaneServiceProvider of(BuildContext context) {
     final ArcaneServiceProvider? result =
         context.dependOnInheritedWidgetOfExactType<ArcaneServiceProvider>();
@@ -29,17 +64,27 @@ class ArcaneServiceProvider extends InheritedNotifier {
   }
 }
 
+/// An extension on `BuildContext` to provide easy access to `ArcaneService` instances
+/// that are registered in an `ArcaneServiceProvider`.
+///
+/// This extension provides a `serviceOfType` method, which searches for a specific
+/// service of type `T` in the current `ArcaneServiceProvider` or in the list of built-in
+/// services.
+///
+/// Example usage:
+/// ```dart
+/// final MyService? myService = context.serviceOfType<MyService>();
+/// ```
 extension ServiceProvider on BuildContext {
-  /// Provides the `serviceOfType` extension on `BuildContext` to find a given
-  /// `ArcaneService` instance that has been registered in the
-  /// `ArcaneServiceProvider`.
+  /// Finds and returns the `ArcaneService` instance of type `T` that has been registered
+  /// in the `ArcaneServiceProvider` or in the list of built-in services (`Arcane.services`).
   ///
-  /// Returns either the requested `ArcaneService.I` or null if one cannot be
-  /// found.
+  /// If no such service is found, it returns `null`.
   ///
-  /// Usage:
+  /// - `T`: The type of the service to be retrieved, which extends `ArcaneService`.
   ///
-  /// ```
+  /// Example:
+  /// ```dart
   /// final MyService? myService = context.serviceOfType<MyService>();
   /// ```
   T? serviceOfType<T extends ArcaneService>() {
@@ -56,4 +101,10 @@ extension ServiceProvider on BuildContext {
   }
 }
 
+/// An abstract class representing a service in the Arcane architecture.
+///
+/// Classes that extend `ArcaneService` can use `ChangeNotifier` functionality
+/// to notify listeners of changes. Services are typically registered in
+/// `ArcaneServiceProvider` and can be accessed using the `serviceOfType`
+/// method on `BuildContext`.
 abstract class ArcaneService with ChangeNotifier {}
